// frontend/src/api/generated/models/personDto.ts
/**
 * Generated by orval v8.3.0 🍺
 * Do not edit manually.
 * Backend API
 * A simple ASP.NET Core Web API template with best practices for building scalable and maintainable applications.
 * OpenAPI spec version: v1
 */
import type { FactoryDto } from "./factoryDto";

export interface PersonDto {
  id: number;
  personalId: string;
  fullName: string;
  email: string;
  allowedFactories: FactoryDto[];
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}